add owner attribute on process struct:
For EVM processes the entityId is a token
contract address. If the process wants to
have the capacity for incrementing a process
question index and given the process having
the required properties enabled to do so, there is
no way to call the incrementQuestionIndex method
for the process because the token contract does not
always have the capacity of calling the method.
Adding the owner attribute enables to increment
the question index by the creator of a process.
This also enables the possibility of changing
some EVM process atritibutes on the future
by just having to check if the msg.sender
is the creator of a process.